move price utils into price crate
