+++ core/Cargo.toml
@@ -33,6 +33,7 @@ log = "0.4"
 chrono = { version = "0.4.4", features = ["serde"] }
 zeroize = "0.9"
 
+diff0 = { path = "../diff0", version = "0.1.0" }
 gotts_keychain = { path = "../keychain", version = "0.0.6" }
 gotts_util = { path = "../util", version = "0.0.6" }
