+++ core/src/core/price.rs
@@ -24,8 +24,13 @@ use crate::libtx::secp_ser;
 use crate::ser::{self, read_multi, Readable, Reader, Writeable, Writer};
 use crate::util::secp::{self, PublicKey, Signature};
 use crate::util::static_secp_instance;
-use crate::util::RwLock;
+use crate::util::{to_hex, RwLock};
+use diff0::{self, diff0_compress, diff0_decompress};
+
 use std::sync::Arc;
+
+/// The price data precision in fraction (1/x)
+pub const GOTTS_PRICE_PRECISION: f64 = 1_000_000_000.0_f64;
 
 /// Data queried for the exchange rate of a currency pair.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -98,7 +103,7 @@ impl ExchangeRates {
 		}
 
 		Ok(ExchangeRates {
-			version: PriceVersion::default(),
+			version: PriceVersion::Raw(0),
 			source_uid: price_feeder_source_uid,
 			pairs,
 			date,
@@ -193,47 +198,98 @@ impl Writeable for ExchangeRates {
 	}
 }
 
+/// Encoded Price Data with PriceVersion
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct VersionedPriceEncoded {
+	/// Price version
+	pub version: PriceVersion,
+	/// Encoded Price Data, variable length vector
+	pub encoded_price: Vec<u8>,
+}
+
+impl Writeable for VersionedPriceEncoded {
+	fn write<W: Writer>(&self, writer: &mut W) -> Result<(), ser::Error> {
+		self.version.write(writer)?;
+		writer.write_u16(self.encoded_price.len() as u16)?;
+		writer.write_fixed_bytes(&self.encoded_price)
+	}
+}
+
+impl Readable for VersionedPriceEncoded {
+	fn read(reader: &mut dyn Reader) -> Result<VersionedPriceEncoded, ser::Error> {
+		let version = PriceVersion::read(reader)?;
+		let encoded_price_len = reader.read_u16()?;
+		let encoded_price = reader.read_fixed_bytes(encoded_price_len as usize)?;
+		Ok(VersionedPriceEncoded {
+			version,
+			encoded_price,
+		})
+	}
+}
+
+impl VersionedPriceEncoded {
+	/// Constructor
+	pub fn new(version: u16, encoded_price: Vec<u8>, is_raw: bool) -> VersionedPriceEncoded {
+		let version = match is_raw {
+			true => PriceVersion::Raw(version),
+			false => PriceVersion::Diff(version),
+		};
+		VersionedPriceEncoded {
+			version,
+			encoded_price,
+		}
+	}
+
+	/// Size of this object
+	pub fn size(&self) -> usize {
+		self.encoded_price.len() + 2
+	}
+}
+
+// The default methold is only for test.
+//impl Default for VersionedPriceEncoded {
+//	fn default() -> VersionedPriceEncoded {
+//		VersionedPriceEncoded::new(0)
+//	}
+//}
+
 /// Some type safety around price versioning.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct PriceVersion(pub u16);
-
-impl Default for PriceVersion {
-	fn default() -> PriceVersion {
-		PriceVersion(0)
-	}
-}
-
-// self-conscious increment function courtesy of Jasper
-impl PriceVersion {
-	#[allow(dead_code)]
-	fn next(&self) -> Self {
-		Self(self.0 + 1)
-	}
-}
-
-impl PriceVersion {
-	/// Constructor taking the provided version.
-	pub fn new(version: u16) -> PriceVersion {
-		PriceVersion(version)
-	}
+pub enum PriceVersion {
+	/// Differential Price
+	Diff(u16),
+	/// Raw Price
+	Raw(u16),
 }
 
 impl From<PriceVersion> for u16 {
-	fn from(v: PriceVersion) -> u16 {
-		v.0
+	fn from(ver: PriceVersion) -> u16 {
+		match ver {
+			PriceVersion::Diff(v) => v | 0x8000u16,
+			PriceVersion::Raw(v) => v & 0x7fffu16,
+		}
+	}
+}
+
+impl From<u16> for PriceVersion {
+	fn from(v: u16) -> PriceVersion {
+		match v >> 15 {
+			0 => PriceVersion::Diff(v & 0x7fffu16),
+			_ => PriceVersion::Raw(v & 0x7fffu16),
+		}
 	}
 }
 
 impl Writeable for PriceVersion {
 	fn write<W: Writer>(&self, writer: &mut W) -> Result<(), ser::Error> {
-		writer.write_u16(self.0)
+		writer.write_u16(u16::from(self.clone()))
 	}
 }
 
 impl Readable for PriceVersion {
 	fn read(reader: &mut dyn Reader) -> Result<PriceVersion, ser::Error> {
 		let version = reader.read_u16()?;
-		Ok(PriceVersion(version))
+		Ok(PriceVersion::from(version))
 	}
 }
 
@@ -296,6 +352,117 @@ impl From<Error> for PoolError {
 	fn from(e: Error) -> PoolError {
 		PoolError::InvalidPrice(e)
 	}
+}
+
+/// Diff0 compression encoder to convert float price into fixed point binary
+pub fn encode_price_feeder(
+	previous_price_pairs: &Option<Vec<f64>>,
+	aggregated_rates: &Vec<ExchangeRate>,
+) -> Vec<u8> {
+	let price_pairs: Vec<f64> = aggregated_rates.iter().map(|r| r.rate).collect();
+	let precision = GOTTS_PRICE_PRECISION;
+
+	let pairs = if let Some(previous) = previous_price_pairs {
+		// only encode the difference values, with 10^-9 precision.
+		assert_eq!(previous.len(), price_pairs.len());
+		let diff: Vec<f64> = price_pairs
+			.iter()
+			.zip(previous)
+			.map(|(a, b)| ((a - b) * precision).round() / precision)
+			.collect();
+		debug!("diff price pairs = {:?}", diff);
+		diff
+	} else {
+		// encode the raw values
+		debug!("raw price pairs = {:?}", price_pairs);
+		price_pairs
+	};
+
+	// convert float64 to i64 with 10^-9 precision
+	let i64_pairs: Vec<i64> = pairs
+		.iter()
+		.map(|r| (*r * precision).round() as i64)
+		.collect();
+
+	// encode
+	let serialized_buffer = diff0_compress(i64_pairs).unwrap();
+	debug!(
+		"serialized_buffer: len = {}, data = {}",
+		serialized_buffer.len(),
+		to_hex(serialized_buffer.clone())
+	);
+	serialized_buffer
+}
+
+/// Util to calculate full price pairs from the basic pairs
+pub fn calculate_full_price_pairs(aggregated_rates: &Vec<ExchangeRate>) -> Result<(), Error> {
+	let currencies_a = vec!["EUR", "GBP", "BTC", "ETH"];
+	let currencies_b = vec!["CNY", "JPY", "CAD"];
+	let mut calculated_rates: Vec<ExchangeRate> = vec![];
+
+	// firstly, get/calculate all x over USD rates
+	for from in currencies_a.clone() {
+		let to = "USD";
+		let index = aggregated_rates
+			.iter()
+			.position(|r| r.from == from && r.to == to)
+			.ok_or(Error::Generic(format!("price {}2{} not found", from, to)))?;
+		calculated_rates.push(aggregated_rates[index].clone());
+	}
+	for to in currencies_b.clone().into_iter() {
+		let index = aggregated_rates
+			.iter()
+			.position(|r| r.from == "USD" && r.to == to)
+			.ok_or(Error::Generic(format!("price USD2{} not found", to)))?;
+		let rate = ExchangeRate {
+			from: to.to_string(),
+			to: "USD".to_string(),
+			rate: 1f64 / aggregated_rates[index].rate,
+			date: aggregated_rates[index].date,
+		};
+		calculated_rates.push(rate);
+	}
+
+	// secondly, calculate/get 1/(x/USD) to get the rates of USD over all x.
+	for (index, to) in currencies_a.iter().enumerate() {
+		let rate = ExchangeRate {
+			from: "USD".to_string(),
+			to: to.to_string(),
+			rate: 1f64 / calculated_rates[index].rate,
+			date: calculated_rates[index].date,
+		};
+		calculated_rates.push(rate);
+	}
+	for to in currencies_b.clone().into_iter() {
+		let index = aggregated_rates
+			.iter()
+			.position(|r| r.from == "USD" && r.to == to)
+			.ok_or(Error::Generic(format!("price USD2{} not found", to)))?;
+		calculated_rates.push(aggregated_rates[index].clone());
+	}
+
+	// thirdly, calculate all others
+	let currencies = [&currencies_a[..], &currencies_b[..]].concat();
+	for (i, from) in currencies.iter().enumerate() {
+		for (j, to) in currencies.iter().enumerate() {
+			if i != j {
+				let rate = ExchangeRate {
+					from: from.to_string(),
+					to: to.to_string(),
+					rate: calculated_rates[i].rate / calculated_rates[j].rate,
+					date: calculated_rates[i].date,
+				};
+				calculated_rates.push(rate);
+			}
+		}
+	}
+
+	trace!(
+		"price pairs = {}",
+		serde_json::to_string_pretty(&calculated_rates).unwrap()
+	);
+
+	Ok(())
 }
 
 #[cfg(test)]
